# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "In Abstract Controller",
        "version": "1"
    },
    "paths": {
...
        "/users/mobile-offers/{id}": {
            "get": {
                "tags": [
                    "usersMobileOffers"
                ],
                "operationId": "3d2e751f79d7da96b480fc5725c000fe",
                "responses": {
                    "200": {
                        "description": "Item fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserMobileOfferGetMobileOfferResponseGetMobileOfferResponseData"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "storesMobileOffers"
                ],
                "operationId": "2d2cc711fb58272f338501af981f9bfa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMobileOfferRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/usersGetMobileOfferRequestPathParameterData"
                }
            ]
        },
...
        "/users/auth/login/login": {
            "post": {
                "tags": [
                    "usersAuth"
                ],
                "operationId": "ef47ca2825fc9ac67f7a6fae28934542",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponseData"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/auth/registeration/phone-number-step": {
            "post": {
                "tags": [
                    "usersAuth"
                ],
                "operationId": "aefe4cf7655ae7cc7f4a98206a67d6c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddPhoneNumberRegisterationStepRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddPhoneNumberRegisterationStepResponseData"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/mobile-offers": {
            "post": {
                "tags": [
                    "usersMobileOffers"
                ],
                "operationId": "efa9d6357f0f68596212f4a73ef4d56b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserMobileOfferCreateMobileOfferRequestCreateMobileOfferRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MediaData": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "file_url": {
                        "type": "string"
                    },
                    "thumbnail_url": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserMobileOfferCreateMobileOfferRequestCreateMobileOfferRequestData": {
                "properties": {
                    "name_in_english": {
                        "type": "string",
                        "nullable": true
                    },
                    "price_in_usd": {
                        "type": "integer",
                        "nullable": true
                    },
                    "ram": {
                        "type": "string",
                        "nullable": true
                    },
                    "storage": {
                        "type": "string",
                        "nullable": true
                    },
                    "battery_health": {
                        "type": "integer",
                        "nullable": true
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserMobileOfferCreateMobileOfferRequestCreateMobileOfferRequestDataFeatureData"
                        },
                        "nullable": true
                    },
                    "temporary_uploaded_images_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserMobileOfferCreateMobileOfferRequestCreateMobileOfferRequestDataFeatureData": {
                "properties": {
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "LoginRequestData": {
                "properties": {
                    "phone_number": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginResponseData": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AddPhoneNumberRegisterationStepRequestData": {
                "properties": {
                    "phone_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AddPhoneNumberRegisterationStepResponseData": {
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserMyMobileOfferGetMobileOfferRequestFeatuerDataGetMobileOfferResponseData": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserMobileOfferGetMobileOfferResponseGetMobileOfferResponseData": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name_in_english": {
                        "type": "string"
                    },
                    "name_in_arabic": {
                        "type": "string"
                    },
                    "price_in_usd": {
                        "type": "integer",
                        "nullable": true
                    },
                    "ram": {
                        "type": "string",
                        "nullable": true
                    },
                    "storage": {
                        "type": "string",
                        "nullable": true
                    },
                    "battery_health": {
                        "type": "integer",
                        "nullable": true
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserMyMobileOfferGetMobileOfferRequestFeatuerDataGetMobileOfferResponseData"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaData"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateMobileOfferRequestData": {
                "properties": {
                    "name_in_english": {
                        "type": "string",
                        "nullable": true
                    },
                    "price_in_usd": {
                        "type": "integer",
                        "nullable": true
                    },
                    "ram": {
                        "type": "string",
                        "nullable": true
                    },
                    "storage": {
                        "type": "string",
                        "nullable": true
                    },
                    "battery_health": {
                        "type": "integer",
                        "nullable": true
                    },
                    "temporary_uploaded_images_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
...
            "usersGetMobileOfferRequestPathParameterData": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ],
...
}